+++ BraintreeCore/src/main/java/com/braintreepayments/api/core/AnalyticsEvent.kt
@@ -1,5 +1,10 @@
 package com.braintreepayments.api.core
 
+/**
+ * DTO for analytics events. See also: [AnalyticsEventParams]
+ * This class is internal to core module and is used in [AnalyticsClient] to construct the analytics
+ * payload to be sent to the backend.
+ */
 internal data class AnalyticsEvent(
     val name: String,
     val timestamp: Long,
@@ -8,5 +13,7 @@ internal data class AnalyticsEvent(
     val isVaultRequest: Boolean = false,
     val startTime: Long? = null,
     val endTime: Long? = null,
-    val endpoint: String? = null
+    val endpoint: String? = null,
+    val experiment: String? = null,
+    val paymentMethodsDisplayed: List<String> = emptyList()
 )